Starting with basic stripboard layout; validated.
Validation is done in VeeCAD using the netlist generated in KiCAD
from the muRata-MZB3004T04-driver.(pro|sch) files.

Before all that, I took the driver schematic in the MZB3004T04 datasheet
and recreated it in KiCAD (being careful to use VeeCAD's libraries (V_*.lib)
of parts so that the resulting netlist would be interpretable by VeeCAD).