Rename 'Vin' to 'Vm'
The designator was taken from the MZB3004T04 datasheet, but
due to low image fidelity, it was hard to tell exactly what
character(s) followed the 'V' for this particular label.

It has occurred to me, of course, that 'Vin' might be a very confusing
interpretation. After all, this is not the input voltage for the board.
That voltage is 'Vs', presumably "voltage source". Given the
resistor-divider circuit that produces 'Vm'/'Vin', I suspect
that this is a half-voltage, so "voltage mid" or "voltage middle"
might be a less confusing interpretation, hence 'Vm'.